# openapi_temp.json
{"openapi":"3.1.0","info":{"title":"AI Assistant API","description":"意图识别和处理API","version":"0.1.0"},"paths":{"/health":{"get":{"tags":["health"],"summary":"健康检查","description":"检查系统运行状态，返回当前时间戳和状态信息。","operationId":"health_check_health_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/v1/intent/interpret":{"post":{"tags":["intent","intent"],"summary":"Process Intent","description":"处理用户意图。根据用户查询，可能返回需要调用的工具列表，或直接返回回复内容。\n\nArgs:\n    request: 意图请求\n    db: 数据库会话 (由 FastAPI 注入)\n    current_user: 当前认证用户 (由JWT令牌提供，可为None)\n\nReturns:\n    意图响应","operationId":"process_intent_api_v1_intent_interpret_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/IntentRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/InterpretSuccessResponse"}}}},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"Bearer认证":[]}]}},"/api/v1/intent/confirm":{"post":{"tags":["intent","intent"],"summary":"Confirm Execution","description":"处理用户确认执行请求。当用户确认执行工具调用时，系统将执行相应的工具并返回结果。\n\nArgs:\n    request: 确认请求\n    db: 数据库会话 (由 FastAPI 注入)\n    current_user: 当前认证用户 (由JWT令牌提供)\n    \nReturns:\n    确认执行响应","operationId":"confirm_execution_api_v1_intent_confirm_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConfirmRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ConfirmResponse"}}}},"404":{"description":"Not found"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"Bearer认证":[]}]}},"/api/v1/execute":{"post":{"tags":["execute"],"summary":"执行指定工具","description":"使用给定参数执行指定工具并返回结果。","operationId":"execute_tool_api_v1_execute_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExecuteRequest"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ExecuteResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"Bearer认证":[]}]}},"/api/v1/tools":{"get":{"tags":["tools"],"summary":"获取可用工具列表","description":"检索系统中所有可用工具的列表。","operationId":"get_tools_api_v1_tools_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ToolsListResponse"}}}}},"security":[{"Bearer认证":[]}]}},"/api/v1/auth/register":{"post":{"tags":["auth","auth"],"summary":"Register","description":"注册新用户\n\nArgs:\n    user_data: 用户创建请求\n    db: 数据库会话\n    \nReturns:\n    新创建的用户信息\n    \nRaises:\n    HTTPException: 如果用户名已存在","operationId":"register_api_v1_auth_register_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"401":{"description":"未授权"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/token":{"post":{"tags":["auth","auth"],"summary":"Login For Access Token","description":"用户登录获取访问令牌\n\nArgs:\n    form_data: 表单数据，包含用户名和密码\n    db: 数据库会话\n    \nReturns:\n    访问令牌和相关信息\n    \nRaises:\n    HTTPException: 如果用户名或密码错误","operationId":"login_for_access_token_api_v1_auth_token_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_for_access_token_api_v1_auth_token_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"401":{"description":"未授权"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/login":{"post":{"tags":["auth","auth"],"summary":"Login","description":"用户登录获取访问令牌（兼容性端点）\n\nArgs:\n    form_data: 表单数据，包含用户名和密码\n    db: 数据库会话\n    \nReturns:\n    访问令牌和相关信息\n    \nRaises:\n    HTTPException: 如果用户名或密码错误","operationId":"login_api_v1_auth_login_post","requestBody":{"content":{"application/x-www-form-urlencoded":{"schema":{"$ref":"#/components/schemas/Body_login_api_v1_auth_login_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/TokenResponse"}}}},"401":{"description":"未授权"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/auth/me":{"get":{"tags":["auth","auth"],"summary":"Get Me","description":"获取当前登录用户信息\n\nArgs:\n    current_user: 当前用户，通过依赖项注入\n    \nReturns:\n    当前用户信息","operationId":"get_me_api_v1_auth_me_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/UserResponse"}}}},"401":{"description":"未授权"}},"security":[{"Bearer认证":[]}]}},"/api/v1/dev/tools":{"get":{"tags":["dev-tools","dev-tools"],"summary":"Get Developer Tools","description":"获取开发者工具列表\n\nArgs:\n    page: 页码\n    page_size: 每页大小\n    status: 工具状态筛选\n    is_public: 是否公开筛选\n    search: 搜索关键词\n    current_user: 当前开发者用户\n    db: 数据库会话\n    \nReturns:\n    开发者工具列表","operationId":"get_developer_tools_api_v1_dev_tools_get","security":[{"Bearer认证":[]}],"parameters":[{"name":"page","in":"query","required":false,"schema":{"type":"integer","minimum":1,"description":"页码","default":1,"title":"Page"},"description":"页码"},{"name":"page_size","in":"query","required":false,"schema":{"type":"integer","maximum":100,"minimum":1,"description":"每页大小","default":10,"title":"Page Size"},"description":"每页大小"},{"name":"status","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"工具状态筛选","title":"Status"},"description":"工具状态筛选"},{"name":"is_public","in":"query","required":false,"schema":{"anyOf":[{"type":"boolean"},{"type":"null"}],"description":"是否公开筛选","title":"Is Public"},"description":"是否公开筛选"},{"name":"search","in":"query","required":false,"schema":{"anyOf":[{"type":"string"},{"type":"null"}],"description":"搜索关键词","title":"Search"},"description":"搜索关键词"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeveloperToolListResponse"}}}},"401":{"description":"未授权"},"403":{"description":"权限不足"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["dev-tools","dev-tools"],"summary":"Create Developer Tool","description":"创建新的开发者工具\n\nArgs:\n    tool_data: 工具创建数据\n    current_user: 当前开发者用户\n    db: 数据库会话\n    \nReturns:\n    创建的工具信息\n    \nRaises:\n    HTTPException: 如果工具ID已存在","operationId":"create_developer_tool_api_v1_dev_tools_post","security":[{"Bearer认证":[]}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeveloperToolCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeveloperToolResponse"}}}},"401":{"description":"未授权"},"403":{"description":"权限不足"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/dev/tools/{tool_id}":{"get":{"tags":["dev-tools","dev-tools"],"summary":"Get Developer Tool","description":"获取开发者工具详情\n\nArgs:\n    tool_id: 工具ID\n    current_user: 当前开发者用户\n    db: 数据库会话\n    \nReturns:\n    工具详情\n    \nRaises:\n    HTTPException: 如果工具不存在或无权限访问","operationId":"get_developer_tool_api_v1_dev_tools__tool_id__get","security":[{"Bearer认证":[]}],"parameters":[{"name":"tool_id","in":"path","required":true,"schema":{"type":"string","title":"Tool Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeveloperToolResponse"}}}},"401":{"description":"未授权"},"403":{"description":"权限不足"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"put":{"tags":["dev-tools","dev-tools"],"summary":"Update Developer Tool","description":"更新开发者工具\n\nArgs:\n    tool_id: 工具ID\n    tool_data: 工具更新数据\n    current_user: 当前开发者用户\n    db: 数据库会话\n    \nReturns:\n    更新后的工具信息\n    \nRaises:\n    HTTPException: 如果工具不存在或无权限访问","operationId":"update_developer_tool_api_v1_dev_tools__tool_id__put","security":[{"Bearer认证":[]}],"parameters":[{"name":"tool_id","in":"path","required":true,"schema":{"type":"string","title":"Tool Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeveloperToolUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DeveloperToolResponse"}}}},"401":{"description":"未授权"},"403":{"description":"权限不足"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"delete":{"tags":["dev-tools","dev-tools"],"summary":"Delete Developer Tool","description":"删除开发者工具\n\nArgs:\n    tool_id: 工具ID\n    current_user: 当前开发者用户\n    db: 数据库会话\n    \nReturns:\n    删除成功消息\n    \nRaises:\n    HTTPException: 如果工具不存在或无权限访问","operationId":"delete_developer_tool_api_v1_dev_tools__tool_id__delete","security":[{"Bearer认证":[]}],"parameters":[{"name":"tool_id","in":"path","required":true,"schema":{"type":"string","title":"Tool Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"401":{"description":"未授权"},"403":{"description":"权限不足"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/dev/upload":{"post":{"tags":["dev-tools","dev-tools"],"summary":"Upload Tool Package","description":"上传工具包文件\n\nArgs:\n    file: 上传的文件\n    current_user: 当前开发者用户\n    db: 数据库会话\n    \nReturns:\n    上传结果","operationId":"upload_tool_package_api_v1_dev_upload_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_upload_tool_package_api_v1_dev_upload_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ToolUploadResponse"}}}},"401":{"description":"未授权"},"403":{"description":"权限不足"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"Bearer认证":[]}]}},"/api/v1/dev/upload/validate":{"post":{"tags":["dev-tools","dev-tools"],"summary":"Validate Tool Package","description":"验证工具包格式\n\nArgs:\n    file: 上传的文件\n    current_user: 当前开发者用户\n    \nReturns:\n    验证结果","operationId":"validate_tool_package_api_v1_dev_upload_validate_post","requestBody":{"content":{"multipart/form-data":{"schema":{"$ref":"#/components/schemas/Body_validate_tool_package_api_v1_dev_upload_validate_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"401":{"description":"未授权"},"403":{"description":"权限不足"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"Bearer认证":[]}]}},"/api/v1/dev/tools/{tool_id}/test":{"post":{"tags":["dev-tools","dev-tools"],"summary":"Test Developer Tool","description":"测试指定的开发者工具\n\nArgs:\n    tool_id: 工具ID\n    test_data: 测试数据\n    current_user: 当前开发者用户\n    db: 数据库会话\n    \nReturns:\n    测试结果","operationId":"test_developer_tool_api_v1_dev_tools__tool_id__test_post","security":[{"Bearer认证":[]}],"parameters":[{"name":"tool_id","in":"path","required":true,"schema":{"type":"string","title":"Tool Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ToolTestRequest"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/ToolTestResponse"}}}},"401":{"description":"未授权"},"403":{"description":"权限不足"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/dev/tools/test/batch":{"post":{"tags":["dev-tools","dev-tools"],"summary":"Batch Test Tools","description":"批量测试工具\n\nArgs:\n    tool_ids: 工具ID列表\n    test_data: 测试数据\n    current_user: 当前开发者用户\n    db: 数据库会话\n    \nReturns:\n    批量测试结果","operationId":"batch_test_tools_api_v1_dev_tools_test_batch_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Body_batch_test_tools_api_v1_dev_tools_test_batch_post"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"401":{"description":"未授权"},"403":{"description":"权限不足"},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"Bearer认证":[]}]}},"/api/v1/mcp/status":{"get":{"tags":["mcp-status"],"summary":"获取所有MCP服务器状态","description":"获取系统中所有MCP服务器的运行状态信息，包括运行状态、重启次数等。","operationId":"get_mcp_servers_status_api_v1_mcp_status_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"additionalProperties":true,"type":"object","title":"Response Get Mcp Servers Status Api V1 Mcp Status Get"}}}}},"security":[{"Bearer认证":[]}]}},"/api/v1/mcp/status/{server_name}":{"get":{"tags":["mcp-status"],"summary":"获取指定MCP服务器状态","description":"获取指定名称的MCP服务器的详细状态信息。","operationId":"get_mcp_server_status_api_v1_mcp_status__server_name__get","security":[{"Bearer认证":[]}],"parameters":[{"name":"server_name","in":"path","required":true,"schema":{"type":"string","title":"Server Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Get Mcp Server Status Api V1 Mcp Status  Server Name  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/mcp/restart/{server_name}":{"post":{"tags":["mcp-status"],"summary":"重启MCP服务器","description":"重启指定的MCP服务器。需要管理员权限。","operationId":"restart_mcp_server_api_v1_mcp_restart__server_name__post","security":[{"Bearer认证":[]}],"parameters":[{"name":"server_name","in":"path","required":true,"schema":{"type":"string","title":"Server Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Restart Mcp Server Api V1 Mcp Restart  Server Name  Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/mcp/start/{server_name}":{"post":{"tags":["mcp-status"],"summary":"启动MCP服务器","description":"启动指定的MCP服务器。需要管理员权限。","operationId":"start_mcp_server_api_v1_mcp_start__server_name__post","security":[{"Bearer认证":[]}],"parameters":[{"name":"server_name","in":"path","required":true,"schema":{"type":"string","title":"Server Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Start Mcp Server Api V1 Mcp Start  Server Name  Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/v1/mcp/stop/{server_name}":{"post":{"tags":["mcp-status"],"summary":"停止MCP服务器","description":"停止指定的MCP服务器。需要管理员权限。","operationId":"stop_mcp_server_api_v1_mcp_stop__server_name__post","security":[{"Bearer认证":[]}],"parameters":[{"name":"server_name","in":"path","required":true,"schema":{"type":"string","title":"Server Name"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Response Stop Mcp Server Api V1 Mcp Stop  Server Name  Post"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"Body_batch_test_tools_api_v1_dev_tools_test_batch_post":{"properties":{"tool_ids":{"items":{"type":"string"},"type":"array","title":"Tool Ids"},"test_data":{"$ref":"#/components/schemas/ToolTestRequest"}},"type":"object","required":["tool_ids","test_data"],"title":"Body_batch_test_tools_api_v1_dev_tools_test_batch_post"},"Body_login_api_v1_auth_login_post":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"^password$"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_login_api_v1_auth_login_post"},"Body_login_for_access_token_api_v1_auth_token_post":{"properties":{"grant_type":{"anyOf":[{"type":"string","pattern":"^password$"},{"type":"null"}],"title":"Grant Type"},"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"},"scope":{"type":"string","title":"Scope","default":""},"client_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Id"},"client_secret":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Client Secret"}},"type":"object","required":["username","password"],"title":"Body_login_for_access_token_api_v1_auth_token_post"},"Body_upload_tool_package_api_v1_dev_upload_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_upload_tool_package_api_v1_dev_upload_post"},"Body_validate_tool_package_api_v1_dev_upload_validate_post":{"properties":{"file":{"type":"string","format":"binary","title":"File"}},"type":"object","required":["file"],"title":"Body_validate_tool_package_api_v1_dev_upload_validate_post"},"ConfirmRequest":{"properties":{"session_id":{"type":"string","title":"Session Id","description":"会话ID"},"user_input":{"type":"string","title":"User Input","description":"用户的自然语言输入"}},"type":"object","required":["session_id","user_input"],"title":"ConfirmRequest","description":"用户确认请求模型"},"ConfirmResponse":{"properties":{"session_id":{"type":"string","title":"Session Id","description":"会话ID"},"success":{"type":"boolean","title":"Success","description":"是否执行成功"},"content":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Content","description":"执行结果内容"},"error":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Error","description":"错误信息"}},"type":"object","required":["session_id","success"],"title":"ConfirmResponse","description":"确认执行响应模型"},"DeveloperToolCreate":{"properties":{"tool_id":{"type":"string","title":"Tool Id","description":"工具唯一标识"},"name":{"type":"string","title":"Name","description":"工具名称"},"type":{"type":"string","title":"Type","description":"工具类型，如 'mcp' 或 'http'"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description","description":"工具描述"},"endpoint":{"additionalProperties":true,"type":"object","title":"Endpoint","description":"工具端点配置"},"request_schema":{"additionalProperties":true,"type":"object","title":"Request Schema","description":"请求参数的JSON Schema"},"response_schema":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Response Schema","description":"响应的JSON Schema"},"server_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Server Name","description":"MCP服务器名称"},"is_public":{"type":"boolean","title":"Is Public","description":"是否公开可用","default":true},"version":{"type":"string","title":"Version","description":"工具版本","default":"1.0.0"},"tags":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Tags","description":"工具标签"}},"type":"object","required":["tool_id","name","type","endpoint","request_schema"],"title":"DeveloperToolCreate","description":"开发者工具创建请求模型"},"DeveloperToolListResponse":{"properties":{"tools":{"items":{"$ref":"#/components/schemas/DeveloperToolResponse"},"type":"array","title":"Tools","description":"工具列表"},"total":{"type":"integer","title":"Total","description":"总数量"},"page":{"type":"integer","title":"Page","description":"当前页码"},"page_size":{"type":"integer","title":"Page Size","description":"每页大小"}},"type":"object","required":["tools","total","page","page_size"],"title":"DeveloperToolListResponse","description":"开发者工具列表响应模型"},"DeveloperToolResponse":{"properties":{"tool_id":{"type":"string","title":"Tool Id","description":"工具唯一标识"},"name":{"type":"string","title":"Name","description":"工具名称"},"type":{"type":"string","title":"Type","description":"工具类型"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description","description":"工具描述"},"endpoint":{"additionalProperties":true,"type":"object","title":"Endpoint","description":"工具端点配置"},"request_schema":{"additionalProperties":true,"type":"object","title":"Request Schema","description":"请求参数的JSON Schema"},"response_schema":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Response Schema","description":"响应的JSON Schema"},"server_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Server Name","description":"MCP服务器名称"},"developer_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Developer Id","description":"开发者用户ID"},"is_public":{"type":"boolean","title":"Is Public","description":"是否公开可用"},"status":{"type":"string","title":"Status","description":"工具状态"},"version":{"type":"string","title":"Version","description":"工具版本"},"tags":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Tags","description":"工具标签"},"download_count":{"type":"integer","title":"Download Count","description":"下载次数"},"rating":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Rating","description":"用户评分"},"created_at":{"type":"string","format":"date-time","title":"Created At","description":"创建时间"},"updated_at":{"type":"string","format":"date-time","title":"Updated At","description":"更新时间"}},"type":"object","required":["tool_id","name","type","endpoint","request_schema","is_public","status","version","download_count","created_at","updated_at"],"title":"DeveloperToolResponse","description":"开发者工具响应模型"},"DeveloperToolUpdate":{"properties":{"name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Name","description":"工具名称"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description","description":"工具描述"},"endpoint":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Endpoint","description":"工具端点配置"},"request_schema":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Request Schema","description":"请求参数的JSON Schema"},"response_schema":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Response Schema","description":"响应的JSON Schema"},"server_name":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Server Name","description":"MCP服务器名称"},"is_public":{"anyOf":[{"type":"boolean"},{"type":"null"}],"title":"Is Public","description":"是否公开可用"},"status":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Status","description":"工具状态"},"version":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Version","description":"工具版本"},"tags":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Tags","description":"工具标签"}},"type":"object","title":"DeveloperToolUpdate","description":"开发者工具更新请求模型"},"ExecuteRequest":{"properties":{"tool_id":{"type":"string","title":"Tool Id","description":"要执行的工具的ID"},"params":{"additionalProperties":true,"type":"object","title":"Params","description":"传递给工具的参数"},"session_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Session Id","description":"当前会话的ID"},"user_id":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"User Id","description":"发起请求的用户ID"}},"type":"object","required":["tool_id","params"],"title":"ExecuteRequest","description":"执行工具的请求模型"},"ExecuteResponse":{"properties":{"tool_id":{"type":"string","title":"Tool Id","description":"被执行的工具ID"},"success":{"type":"boolean","title":"Success","description":"工具是否成功执行"},"data":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Data","description":"工具执行成功时返回的数据"},"error":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Error","description":"工具执行失败时的错误信息, e.g., {'code': '...', 'message': '...'}"},"session_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Session Id","description":"当前会话的ID"}},"type":"object","required":["tool_id","success"],"title":"ExecuteResponse","description":"执行工具的响应模型"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"IntentRequest":{"properties":{"query":{"type":"string","title":"Query","description":"用户查询"},"session_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Session Id","description":"会话ID"},"user_id":{"type":"integer","title":"User Id","description":"用户ID"},"context":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Context","description":"上下文信息"}},"type":"object","required":["query","user_id"],"title":"IntentRequest","description":"意图请求模型"},"InterpretSuccessResponse":{"properties":{"session_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Session Id","description":"会话ID"},"type":{"type":"string","enum":["tool_call","direct_response"],"title":"Type"},"tool_calls":{"anyOf":[{"items":{"$ref":"#/components/schemas/ToolCall"},"type":"array"},{"type":"null"}],"title":"Tool Calls"},"confirm_text":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Confirm Text"},"content":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Content"}},"type":"object","required":["type"],"title":"InterpretSuccessResponse","description":"统一响应模型，包含所有可能的响应类型"},"TokenResponse":{"properties":{"access_token":{"type":"string","title":"Access Token"},"token_type":{"type":"string","title":"Token Type"},"expires_in":{"type":"integer","title":"Expires In"},"user_id":{"type":"integer","title":"User Id"},"username":{"type":"string","title":"Username"},"role":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Role"}},"type":"object","required":["access_token","token_type","expires_in","user_id","username"],"title":"TokenResponse"},"ToolCall":{"properties":{"tool_id":{"type":"string","title":"Tool Id","description":"要调用的工具ID"},"parameters":{"additionalProperties":true,"type":"object","title":"Parameters","description":"调用工具所需的参数"}},"type":"object","required":["tool_id","parameters"],"title":"ToolCall","description":"单个工具调用信息"},"ToolItem":{"properties":{"tool_id":{"type":"string","title":"Tool Id","description":"工具唯一标识"},"name":{"type":"string","title":"Name","description":"工具名称"},"type":{"type":"string","title":"Type","description":"工具类型，如 'mcp' 或 'http'"},"description":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Description","description":"工具描述"},"source":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Source","description":"工具来源，如 'MCP' 或 '开发者'"}},"type":"object","required":["tool_id","name","type"],"title":"ToolItem","description":"单个工具信息"},"ToolTestRequest":{"properties":{"tool_id":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Tool Id","description":"工具ID（测试已保存工具）"},"tool_config":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Tool Config","description":"工具配置（测试未保存工具）"},"test_data":{"additionalProperties":true,"type":"object","title":"Test Data","description":"测试数据"},"timeout":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Timeout","description":"超时时间（秒）","default":30}},"type":"object","required":["test_data"],"title":"ToolTestRequest","description":"工具测试请求模型"},"ToolTestResponse":{"properties":{"success":{"type":"boolean","title":"Success","description":"测试是否成功"},"result":{"anyOf":[{"additionalProperties":true,"type":"object"},{"type":"null"}],"title":"Result","description":"测试结果"},"error":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Error","description":"错误信息"},"execution_time":{"type":"number","title":"Execution Time","description":"执行时间（秒）"},"timestamp":{"type":"string","format":"date-time","title":"Timestamp","description":"测试时间"}},"type":"object","required":["success","execution_time","timestamp"],"title":"ToolTestResponse","description":"工具测试响应模型"},"ToolUploadResponse":{"properties":{"upload_id":{"type":"string","title":"Upload Id","description":"上传ID"},"status":{"type":"string","title":"Status","description":"上传状态"},"message":{"type":"string","title":"Message","description":"状态消息"},"tools_created":{"anyOf":[{"items":{"type":"string"},"type":"array"},{"type":"null"}],"title":"Tools Created","description":"创建的工具ID列表"}},"type":"object","required":["upload_id","status","message"],"title":"ToolUploadResponse","description":"工具包上传响应模型"},"ToolsListResponse":{"properties":{"tools":{"items":{"$ref":"#/components/schemas/ToolItem"},"type":"array","title":"Tools","description":"工具列表"}},"type":"object","required":["tools"],"title":"ToolsListResponse","description":"工具列表响应模型"},"UserCreate":{"properties":{"username":{"type":"string","title":"Username"},"password":{"type":"string","title":"Password"},"email":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Email"}},"type":"object","required":["username","password"],"title":"UserCreate"},"UserResponse":{"properties":{"id":{"type":"integer","title":"Id"},"username":{"type":"string","title":"Username"},"role":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Role"}},"type":"object","required":["id","username"],"title":"UserResponse"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"}},"securitySchemes":{"Bearer认证":{"type":"oauth2","description":"Bearer令牌认证。请在此处输入您的访问令牌。","flows":{"password":{"scopes":{},"tokenUrl":"/api/v1/auth/token"}}}}}}